# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "ServiceBuyer API - сервис для управления подписками пользователей",
        "title": "ServiceBuyer API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/count": {
            "get": {
                "description": "Получить статистику по подпискам с возможностью фильтрации",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "Получить статистику подписок",
                "parameters": [
                    {
                        "description": "Параметры фильтрации",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hirogava_ServiceBuyer_internal_model_request.CountingRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное получение статистики",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Ошибка в данных запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Записи не найдены",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/record": {
            "post": {
                "description": "Создать новую запись о подписке пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "subscriptions"
                ],
                "summary": "Создать подписку",
                "parameters": [
                    {
                        "description": "Данные подписки",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_Hirogava_ServiceBuyer_internal_model_request.ServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешное создание подписки",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка в данных запроса",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Hirogava_ServiceBuyer_internal_model_request.CountingRequest": {
            "type": "object",
            "properties": {
                "end_date": {
                    "type": "string"
                },
                "service_name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "github_com_Hirogava_ServiceBuyer_internal_model_request.ServiceRequest": {
            "type": "object",
            "properties": {
                "cost": {
                    "type": "number"
                },
                "end_date": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "start_date": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        }
    }
}